backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0,<5.0
authlib>=1.4.0
httpx>=0.28.0
apscheduler>=3.10.4
pywebpush>=2.0.0
py-vapid>=1.9.0
aiosmtplib>=3.0.0
jinja2>=3.1.4
openpyxl>=3.1.5
weasyprint>=63.0
python-multipart>=0.0.18
itsdangerous>=2.2.0
email-validator>=2.2.0